system_keyspace: Upgrade compaction_history table
Currently, the system.compaction_history table miss precious
information like the type of compaction (cleanup, major, resharding,
etc) or the sstable generations involved (in and out) used countless
times to diagnose issues.

Thus, the commit extend the current definition of the table by adding
the following columns:
+ "compaction_type" (text)
+ "shard_id" (int16)
+ "sstables_in" (list<sstableinfo_type>)
+ "sstables_out" (list<sstableinfo_type>)
+ "total_tombstone_purge_attempt" (long)
+ "total_tombstone_purge_failure_due_to_overlapping_with_memtable" (long)
+ "total_tombstone_purge_failure_due_to_overlapping_with_uncompacting_sstable" (long)

Furthermore, the commit introduces a new feature flag in order to
prevent nodes from writing data to new columns when a cluster is
not fully upgraded.

Additionally, Unlike user tables -- whose version is generated from
the schema content, including the timestamp the schema is created at
-- system tables have a stable version, generated from the name of
the table. This is to avoid nodes creating the internal table, all
ending up with a different version. When a system table is changed,
its version will not change -- so other nodes will not notice that
there is a newer one. Thus in order to trigger a version change
manually the offset is needed.