database: Add a switch to reject all user table writes
When the switch is on, all mutation writes (includes inserts, updates,
deletes, counter updates, hints) to user tables are rejected, with
a clear error exception returned.

The switch is meant to be used with the disk space monitor in order
to prevent any user writes when node's disk utilization is too high.